# Update the profiling format for Acc32 Benchmark