Summary:
Before this DIFF:
  M, N, K, Packing (ms), Kernel (ms), Postprocessing (ms), Total (ms), GOPs
  3136,    256,     64,         MKL_fp32,  64.5
  0.1, 1.3, 0.3, 1.8,   3136,    256,     64,  FBGEMM_i8_acc32,  55.7

  3136,     64,     64,         MKL_fp32,  54.9
  0.1, 0.3, 0.1, 0.5,   3136,     64,     64,  FBGEMM_i8_acc32,  50.7

  3136,     64,    576,         MKL_fp32,  60.9
  0.4, 2.7, 0.1, 3.3,   3136,     64,    576,  FBGEMM_i8_acc32,  70.3

  ...

After this DIFF:
  M, N, K, Packing (ms), Kernel (ms), Postprocessing (ms), Total (ms), GOPs
  3136,    256,     64,         MKL_fp32,  62.4
  3136,    256,     64, 0.1, 1.3, 0.3, 1.8,  FBGEMM_i8_acc32,  54.8

  3136,     64,     64,         MKL_fp32,  49.4
  3136,     64,     64, 0.1, 0.3, 0.1, 0.5,  FBGEMM_i8_acc32,  46.3

  3136,     64,    576,         MKL_fp32,  65.6
  3136,     64,    576, 0.4, 2.7, 0.1, 3.3,  FBGEMM_i8_acc32,  70.0

  ...

Reviewed By: dskhudia

Differential Revision: D13531989